Added "auto_prune" option to WvStreamList.  The default is 'true', which is the old behaviour of automatically removing !isok() streams from the list. If auto_prune==false, we will not do this, so it's easier to track dying streams.
git-svn-id: file:///home/apenwarr/alumnit-svn/public/trunk@90 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ streams/wvstreamlist.cc
@@ -10,6 +10,7 @@
 
 WvStreamList::WvStreamList()
 {
+    auto_prune = true;
 }
 
 
@@ -33,7 +34,8 @@ bool WvStreamList::select_setup(SelectInfo &si)
 	if (!s.isok())
 	{
 	    one_dead = true;
-	    i.unlink();
+	    if (auto_prune)
+		i.unlink();
 	    continue;
 	}
